rename : apiからfetcherにディレクトリ名を変更
